+++ test/faker/address_test.exs
@@ -7,6 +7,7 @@ defmodule Faker.AddressTest do
   doctest Faker.Address.Hy
   doctest Faker.Address.It
   doctest Faker.Address.Nl
+  doctest Faker.Address.NlBe
   doctest Faker.Address.PtBr
   doctest Faker.Address.Ru
 end